rename main to part1
