+++ 07/part1.py
